wip world.py
new module, took classes out of main.py, wip adding world objects
