+++ game/assets/sprites/tiles/grass_solid.txt
@@ -1,0 +1,8 @@
+░░░░░░░░░░░░░░░░
+░░░░░░░░░░░░░░░░
+░░░░░░░░░░░░░░░░
+░░░░░░░░░░░░░░░░
+░░░░░░░░░░░░░░░░
+░░░░░░░░░░░░░░░░
+░░░░░░░░░░░░░░░░
+░░░░░░░░░░░░░░░░
