+++ game/main.py
@@ -13,154 +13,11 @@ import logging
 import random
 import time
 from typing import NamedTuple
-from common import Character, DialogLine, load_text, load_text_dir, move_toward, remap_dict
+from common import Character, DialogLine, load_text, load_text_dir, remap_dict
 import cuinter
+import world
 
 FPS_COUNTER_REFRESH = 1 # Time between each FPS counter update
-TILE_HEIGHT = 8
-TILE_WIDTH = 16
-TILE_NAME_TO_CHAR = {
-    "blank": "X",
-    "grass": " ",
-    "path_vertical": "│",
-    "path_horizontal": "─",
-    "path_up_left": "┘",
-    "path_up_right": "└",
-    "path_down_left": "┐",
-    "path_down_right": "┌",
-    "wall_vertical": "║",
-    "wall_horizontal": "═",
-    "wall_up_left": "╝",
-    "wall_up_right": "╚",
-    "wall_down_left": "╗",
-    "wall_down_right": "╔",
-}
-WALKABLE_TILE_CHARS = " │─┘└┐┌"
-
-class Player(NamedTuple): # Rename to WorldPlayer or smth if needed
-    character: Character
-    sprite_renderer: cuinter.SpriteRenderer
-    sprite_key: str = "down"
-    
-    @property
-    def y(self) -> int:
-        return self.sprite_renderer.y
-    
-    @property
-    def x(self) -> int:
-        return self.sprite_renderer.x
-    
-    @property
-    def grid_y(self) -> int:
-        return grid.screen_to_grid(self.y, self.x)[0]
-    
-    @property
-    def grid_x(self) -> int:
-        return grid.screen_to_grid(self.y, self.x)[1]
-    
-    @classmethod
-    def new(cls, y: int, x: int, character: Character, sprite_key: str = "down") -> Player:
-        return cls(
-            character,
-            cuinter.SpriteRenderer.new(y, x, character.sprite_sheet[sprite_key]),
-            sprite_key,
-        )
-    
-    def config(self, **kwargs) -> Player:
-        character = kwargs.get("character", self.character)
-        sprite_key = kwargs.get("sprite_key", self.sprite_key)
-        
-        if (sprite_key is not self.sprite_key
-        or character.sprite_sheet is not self.character.sprite_sheet):
-            sprite = character.sprite_sheet[sprite_key]
-        else:
-            sprite = self.sprite_renderer.sprite
-        
-        return Player(
-            character, # TODO allow cross Romain-Jakub config
-            self.sprite_renderer.config(sprite=sprite, **kwargs),
-            sprite_key,
-        )
-    
-    def move(self, y: int, x: int) -> Player:
-        new_grid_y, new_grid_x = self.grid_y + y, self.grid_x + x
-        if not grid.is_walkable(new_grid_y, new_grid_x): return self
-        new_y, new_x = grid.grid_to_screen(new_grid_y, new_grid_x)
-        return self.config(y=new_y, x=new_x)
-
-
-class Grid(NamedTuple):
-    sprite_renderer: cuinter.SpriteRenderer
-    tilemap: tuple[str] # Each str is a row and each char represents a tile
-    tileset: dict[str, str] # Maps the chars in tilemap to their text sprites
-    
-    @property
-    def y(self) -> int:
-        return self.sprite_renderer.y
-    
-    @property
-    def x(self) -> int:
-        return self.sprite_renderer.x
-    
-    @staticmethod
-    def tilemap_to_sprite(tilemap: tuple[str], tileset: dict[str, str]) -> str:
-        """
-        Returns a text sprite of the tilemap with each char replaced by its
-        corresponding tile sprite given by the tileset.
-        """
-        full_sprite = ""
-        for row in tilemap:
-            for sprite_y in range(TILE_HEIGHT):
-                for char in row:
-                    full_sprite += tileset[char].split("\n")[sprite_y]
-                if sprite_y != TILE_HEIGHT - 1 or row != tilemap[-1]:
-                    full_sprite += "\n"
-        return full_sprite
-    
-    @classmethod
-    def new(cls, y: int, x: int, tilemap: tuple[str] = None, tileset: dict[str, str] = None) -> Grid:
-        return cls(
-            cuinter.SpriteRenderer.new(
-                y,
-                x,
-                Grid.tilemap_to_sprite(tilemap, tileset),
-            ),
-            tilemap,
-            tileset,
-        )
-    
-    def config(self, **kwargs) -> Grid:
-        tilemap = kwargs.get("tilemap", self.tilemap)
-        tileset = kwargs.get("tileset", self.tileset)
-        
-        if tilemap is not self.tilemap or tileset is not self.tileset:
-            sprite = Grid.tilemap_to_sprite(tilemap, tileset)
-        else:
-            sprite = self.sprite_renderer.sprite
-        
-        return Grid(
-            self.sprite_renderer.config(sprite=sprite, **kwargs),
-            tilemap, # TODO should update sprite based on new tilemap and tileset
-            tileset,
-        )
-    
-    def grid_to_screen(self, y: int, x: int) -> tuple[int, int]:
-        return (
-            self.y + TILE_HEIGHT * y,
-            self.x + TILE_WIDTH * x,
-        )
-    
-    def screen_to_grid(self, y: int, x: int) -> tuple[int, int]:
-        return (
-            round((y - self.y) / TILE_HEIGHT),
-            round((x - self.x) / TILE_WIDTH),
-        )
-    
-    def is_walkable(self, y: int, x: int) -> bool:
-        if not (0 <= y < len(self.tilemap) and 0 <= x < len(self.tilemap[y])):
-            return False
-        return self.tilemap[y][x] in WALKABLE_TILE_CHARS
-
 
 ############ Code to run on startup
 
@@ -171,16 +28,16 @@ last_time = time.time()
 fps_timer = last_time  # Time of the last FPS update
 frame_count = 0
 
-tilemap = tuple(load_text("assets\\sprites\\tilemaps\\test_tilemap.txt").split("\n"))
-tileset = remap_dict(load_text_dir("assets\\sprites\\tiles"), TILE_NAME_TO_CHAR)
-grid = Grid.new(
-    round((cuinter.screen_height - len(tilemap) * TILE_HEIGHT) / 2),
-    round((cuinter.screen_width - len(tilemap[0]) * TILE_WIDTH) / 2),
-    tilemap,
-    tileset,
-)
+tiles = load_text_dir("assets\\sprites\\tiles")
+tileset = remap_dict(tiles, world.TILE_NAME_TO_CHAR)
+text_tilemap = load_text("assets\\sprites\\tilemaps\\test_tilemap.txt")
+tilemap = tuple(text_tilemap.split("\n"))
 
-player = Player.new(
+grid = world.Grid.new(tileset)
+grid = grid.load_zone(world.Zone(tilemap))
+grid = grid.center(cuinter.screen_height, cuinter.screen_width)
+
+player = world.Player.new(
     grid.y,
     grid.x,
     Character("Player", load_text_dir("assets\\sprites\\characters\\player")),
@@ -254,22 +111,22 @@ while 1:
                 # a UI element
                 # Match case for the key doesn't work, don't waste your time
                 if value == ord("w"):
-                    player = player.move(-1, 0)
+                    player = player.move(grid, -1, 0)
                     if player.sprite_key != "up":
                         player = player.config(sprite_key="up")
                 
                 elif value == ord("s"):
-                    player = player.move(1, 0)
+                    player = player.move(grid, 1, 0)
                     if player.sprite_key != "down":
                         player = player.config(sprite_key="down")
                 
                 elif value == ord("a"):
-                    player = player.move(0, -1)
+                    player = player.move(grid, 0, -1)
                     if player.sprite_key != "left":
                         player = player.config(sprite_key="left")
                 
                 elif value == ord("d"):
-                    player = player.move(0, 1)
+                    player = player.move(grid, 0, 1)
                     if player.sprite_key != "right":
                         player = player.config(sprite_key="right")
